+++ src/parser/part/body.rs
@@ -138,6 +138,10 @@ fn part_def_body_element(input: Input<'_>) -> IResult<Input<'_>, Node<PartDefBod
                 crate::parser::metadata_annotation::metadata_keyword_usage,
                 PartDefBodyElement::MetadataKeywordUsage,
             ),
+            map(
+                crate::parser::metadata_annotation::metadata_annotation,
+                PartDefBodyElement::MetadataAnnotation,
+            ),
             map(annotation, PartDefBodyElement::Annotation),
             map(exhibit_state, PartDefBodyElement::ExhibitState),
             map(calc_usage, PartDefBodyElement::CalcUsage),